WIP: Refactoring code, renaming and decoupling bigger functions into smaller parts
